+++ SolastaCommunityExpansion/Models/_Global.cs
@@ -10,7 +10,8 @@ namespace SolastaCommunityExpansion.Models
         // holds the active player character when in battle
         public static GameLocationCharacter ActivePlayerCharacter { get; set; }
         
-        public static RulesetCharacter InspectedHero { get; set; }
+        // inspected hero on both location and pool
+        public static RulesetCharacterHero InspectedHero { get; set; }
 
         // holds the current action from any character on the map
         public static CharacterAction CurrentAction { get; set; }
